% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziaq.R
\name{ziaq}
\alias{ziaq}
\title{Zero-inflation adjusted quantile regression for single cell RNA sequencing data}
\usage{
ziaq(Y_matrix, colDat, formula = ~condition, group = "condition",
  probs = c(0.25, 0.5, 0.75), log_i = T, parallel = T,
  no.core = detectCores())
}
\arguments{
\item{Y_matrix}{a matrix for expression values with row representing indiviudal genes
and column representing cells}

\item{colDat}{a dataframe including the individual cell information}

\item{formula}{a formula with the predictors included in \code{colDat}. The default
is ~condition.}

\item{group}{the variable name in \code{colDat} for the factor used in group comparsion. The default
is condition.}

\item{probs}{the quantile levels for the quantile regressions.
The default is \code{c(0.25, 0.5, 0.75)}}

\item{log_i}{TRUE or FALSE indicate whether to apply log transformation.
The default is TRUE.}

\item{parallel}{TRUE or FALSE indicate whether to  apply parallel computing.
The default is TRUE.}

\item{no.core}{The number of cores used in parallel computing. The default
is all available cores \code{detectCores()}}
}
\value{
\item{pvalue}{The p-values of all genes for testing the signficance
of the specified \code{group} variable.}
\item{res}{The full results from function \code{ziaq_fit} for all genes}
}
\description{
This function fits the zero-inflation adjusted quantile regression model for the
differential expression analysis in single cell RNA sequencing data
}
\examples{
#Use simuluated data
ymatrix = matrix(round(100* runif(100*500)), ncol = 100)
rownames(ymatrix) = paste0('gene', 1:500)

colDat = data.frame(condition = rep(c(1, 0), e = 50))

res = ziaq(ymatrix, colDat, formula = ~ condition,
          group = 'condition', probs = c(0.25, 0.5, 0.75),
          log_i = T, parallel = T, no.core = 1)

print(res$pvalue)
}
\keyword{ziaq_fit}
