% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziaq_fit.R
\name{ziaq_fit}
\alias{ziaq_fit}
\title{Zero-inflation adjusted quantile regression}
\usage{
ziaq_fit(
  y,
  colDat,
  formula = ~condition,
  group = "condition",
  probs = c(0.25, 0.5, 0.75),
  log_i = T
)
}
\arguments{
\item{y}{a numeric vector including gene expression value for one gen}

\item{colDat}{a dataframe including the individual cell information}

\item{formula}{a formula with the predictors included in \code{colDat}. The default
is ~condition.}

\item{group}{the variable name in \code{colDat} for the factor used in group comparsion. he default
is condition.}

\item{probs}{the quantile levels for the quantile regressions.
The default is \code{c(0.25, 0.5, 0.75)}}

\item{log_i}{TRUE or FALSE indicate whether to apply log transformation.
The default is TRUE.}
}
\value{
\item{pvalue}{The p-value for testing the signficance of the specified \code{group} variable.}
\item{logistic_model}{The \code{glm} object of the logistics regression for fitting binary
outcome (zero or not zero) vs the predictors specified in \code{formula}.}
\item{zero_proportion}{The proportion of zero measurements}
\item{quantile_levels}{The input quantile levels and zero-proportion adjusted quantile levels}
\item{qr}{The objects of the fitted quantile regressions}
}
\description{
This function fits the zero-inflation adjusted quantile regression model
for one individual gene.
}
\examples{
# simulated data
y = round(100* runif(100))
colDat = data.frame(condition = rep(c(1, 0), e = 50))
res = ziaq_fit(y, colDat = colDat,  formula = ~ condition,
              group = 'condition', probs = c(0.25, 0.5, 0.75),
              log_i = TRUE)
print(res)
}
\keyword{ziaq_fit}
